Magnifying lens / Magnifying glass inconsistency fix
The "Magnifying Lens" tool is referred to as "Magnifying Glass" in
several places in the code. As of SVN version 10 it figures in the
toolbar menu as "Magnifying _lens" while the keyboard shortcut is the G
key. This patch replaces "glass" with "lens" in all instances, and fixes
the keyboard shortcuts.

Now the L key, previously unused, activates the lens, while the G key
opens the "Go To Page..." dialog instead of [CTRL+G].

Author: Nephiel <nephiel@users.sourceforge.net>

git-svn-id: svn+ssh://svn.code.sf.net/p/mcomix/code@13 ef8490bc-f317-4725-81ed-e95a9aca1bcf