Provide all menu options within the right-click popup menu
This patch puts all the items of the menubar within the right-click
context menu, neatly organized in submenus. Now every item can be
reached without leaving fullscreen.

Note that everything is also still available in the menubar, with the
only exception of the Bookmarks menu. I still haven't found a good way
to include the bookmarks menu both in the menubar and the right-click
menu... I settled for placing it in the right-click menu, so they are
accessible while in fullscreen.

Author: Nephiel <nephiel@users.sourceforge.net>

git-svn-id: svn://svn.code.sf.net/p/mcomix/code@25 ef8490bc-f317-4725-81ed-e95a9aca1bcf