Add callback decorator.
This class can be wrapped around any function/method
as decorator. Afterwards, the function will always be executed
in the main thread, that is, if another thread calls it,
it will first be queued in Gobject's main loop before
being actually run.

Other than that, wrapped functions can also serve as targets
for the Observer pattern. Add a function with method += callback.
This callback will now be called each time the original function
has been called.

git-svn-id: svn://svn.code.sf.net/p/mcomix/code@171 ef8490bc-f317-4725-81ed-e95a9aca1bcf