Reverted use of GTK stock translations for menu items.
Apparently, there is no way to convince GTK to change its language
once the program has started. It only respects the LANG environment
variable, which cannot be changed once the program is running.

Thus, when changing MComix' language, the stock menu items would
be kept in the original language, which adds a great amount of
inconsistency.

git-svn-id: svn://svn.code.sf.net/p/mcomix/code@308 ef8490bc-f317-4725-81ed-e95a9aca1bcf