Allow recursively adding files to the library.
Selecting a folder will add all file within that folder. Adding should
also be a lot faster now, as thumbnails are no longer generated at add
time, but only when the file is displayed for the first time in the
library.

Since thumbnail generation is done in multiple threads in the display
area, this change should result in a speed increase when adding many books.

git-svn-id: svn+ssh://svn.code.sf.net/p/mcomix/code@359 ef8490bc-f317-4725-81ed-e95a9aca1bcf