Fixed race condition in image handler pre-caching code.
Deleting a file closes the imagehandler, causing the pixbuf cache to
be invalidated. It did not, however, stop the pixbuf cache thread,
which could still add pixbufs even after the image handler was closed.
This sometimes caused MComix to believe that a pixbuf was already
loaded, even though the file indices had already changed. This results
in wrong pixbufs being loaded for certain files.

git-svn-id: svn+ssh://svn.code.sf.net/p/mcomix/code@635 ef8490bc-f317-4725-81ed-e95a9aca1bcf