Remove scroll protection is scrolling is currently impossible.
When an image is smaller than the current view port, there is no need to
scroll the page, however, the scroll protection is still triggered.
This way, the scroll key must be used X times before the image is flipped
to the next page, e.g. in best fit mode.

Fix this by checking if the page can be scrolled at all (ie. image_height >
screen_height), and disable scroll protection if not.

git-svn-id: svn+ssh://svn.code.sf.net/p/mcomix/code@737 ef8490bc-f317-4725-81ed-e95a9aca1bcf